+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 * Update to OKHttp v3.0.1.
 * Add support for prepend user property operation.
 * Add support setting groups for users and events.
-* Migrate shared preferences to Sqlite db to support apps with multiple processes.
+* Migrate shared preferences (userId and event meta data) to Sqlite db to support apps with multiple processes.
 
 ## 2.5.1 (March 14, 2016)